refactor: replace imports from barrel files in an effort to avoid HMR hang in storybook
