+++ lib/components/Meta/Meta.tsx
@@ -2,8 +2,8 @@ import { IconType } from '@autoguru/icons';
 import * as React from 'react';
 import { ComponentProps, FunctionComponent } from 'react';
 
-import { Inline } from '..';
 import { Icon } from '../Icon';
+import { Inline } from '../Inline';
 import { Text } from '../Text';
 
 import * as styles from './Meta.css';
