+++ package.json
@@ -91,7 +91,6 @@
 		"@babel/runtime-corejs3": "7.22.5",
 		"@changesets/cli": "2.26.1",
 		"@chromatic-com/storybook": "^3",
-		"@manypkg/cli": "0.20.0",
 		"@octokit/rest": "18.12.0",
 		"@popperjs/core": "2.11.6",
 		"@react-stately/toggle": "3.7.7",
